feat: improve input field UX and fix all clippy warnings
Input field improvements:
- Remove "Озвучить" button, use Enter to send text
- Shift+Enter for new line (default behavior)
- Clear input immediately after sending (before playback)
- Live replacement on space: \word and %username patterns
- Add load_preprocessor_data command for UI replacements

Backend changes:
- Add PreprocessorData struct and load_preprocessor_data command
- Add get_replacements_map() and get_usernames_map() to TextPreprocessor

Clippy warnings fixed (48 → 0):
- Remove unnecessary casts, redundant field names
- Replace manual Default impls with derive(Default)
- Fix ptr_arg warnings (&PathBuf → &Path)
- Fix enum variant names (remove Invalid prefix)
- Add #[allow] for complex collapsible_match cases
- Fix if_same_then_else and single_match warnings
- Remove empty lines after doc comments

+++ src-tauri/src/commands/mod.rs
@@ -30,8 +30,8 @@ pub fn quit_app(app_handle: AppHandle) -> Result<(), String> {
         // Сохраняем позицию главного окна
         if let Some(main_window) = app_handle.get_webview_window("main") {
             if let Ok(pos) = main_window.outer_position() {
-                let x = pos.x as i32;
-                let y = pos.y as i32;
+                let x = pos.x;
+                let y = pos.y;
                 eprintln!("[APP] Saving main window position: {}, {}", x, y);
                 let _ = windows_manager.set_main_position(Some(x), Some(y));
             }
@@ -41,8 +41,8 @@ pub fn quit_app(app_handle: AppHandle) -> Result<(), String> {
         if let Some(floating_window) = app_handle.get_webview_window("floating") {
             if let Ok(true) = floating_window.is_visible() {
                 if let Ok(pos) = floating_window.outer_position() {
-                    let x = pos.x as i32;
-                    let y = pos.y as i32;
+                    let x = pos.x;
+                    let y = pos.y;
                     eprintln!("[APP] Saving floating window position: {}, {}", x, y);
                     let _ = windows_manager.set_floating_position(Some(x), Some(y));
                 }
